# swagger.json
{"swagger":"2.0","info":{"version":"2.5.0","title":""},"host":"localhost:8080","basePath":"/EHealthCare/webapi","tags":[{"name":"Request Services"},{"name":"Asset Services"},{"name":"Station Services"}],"schemes":["http"],"paths":{"/assetservice":{"post":{"tags":["Asset Services"],"summary":"Adds new asset","description":"","operationId":"addNewAsset","parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"string"},{"name":"np","in":"query","description":"Root ID","required":false,"type":"string"},{"name":"parent","in":"query","description":"Parent ID","required":false,"type":"string"},{"name":"quantity","in":"query","description":"Quantity ID","required":false,"type":"integer","format":"int32"},{"name":"assetType","in":"query","description":"Asset Type ID","required":false,"type":"string"},{"in":"body","name":"body","description":"Asset","required":false,"schema":{"$ref":"#/definitions/NimhansAsset"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/assetservice/addblock":{"post":{"tags":["Asset Services"],"summary":"Adds block for a given tissue","description":"","operationId":"addBlock","parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"string"},{"in":"body","name":"body","description":"Asset","required":false,"schema":{"$ref":"#/definitions/NimhansAsset"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/assetservice/addslide":{"post":{"tags":["Asset Services"],"summary":"Adds slides for a given block","description":"","operationId":"addSlide","parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"string"},{"name":"quantity","in":"query","description":"Quantity","required":false,"type":"string"},{"in":"body","name":"body","description":"Asset","required":false,"schema":{"$ref":"#/definitions/NimhansAsset"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/assetservice/completeblock":{"post":{"tags":["Asset Services"],"summary":"To update the status of a block","description":"","operationId":"completeBlock","parameters":[{"name":"assetId","in":"query","description":"Asset ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/assetservice/completeslide":{"post":{"tags":["Asset Services"],"summary":"To update the status on the given slide","description":"","operationId":"completeSlide","parameters":[{"name":"assetId","in":"query","description":"Asset ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/assetservice/deleteasset":{"delete":{"tags":["Asset Services"],"summary":"Deletes a given asset","description":"","operationId":"deleteAsset","parameters":[{"name":"assetId","in":"query","description":"Asset ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/assetservice/getassets":{"get":{"tags":["Asset Services"],"summary":"Returns all assets of a given Root ID","description":"","operationId":"getAssets","produces":["application/json"],"parameters":[{"name":"npBase","in":"query","description":"Root ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"type":"object"}}}}}},"/assetservice/getchildrenstatus":{"get":{"tags":["Asset Services"],"summary":"Returns the childern's status for a given asset","description":"","operationId":"getChildrenStatus","parameters":[{"name":"assetId","in":"query","description":"Asset ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}}}}},"/assetservice/search":{"get":{"tags":["Asset Services"],"summary":"Search based on filters","description":"","operationId":"search","parameters":[{"name":"rootId","in":"query","description":"Root ID","required":false,"type":"string"},{"name":"requestId","in":"query","description":"Request ID","required":false,"type":"string"},{"name":"patientUHID","in":"query","description":"Patient's UHID","required":false,"type":"string"},{"name":"patientName","in":"query","description":"Patient's Name","required":false,"type":"string"},{"name":"doctorName","in":"query","description":"Doctor's Name","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}},"400":{"description":"No assets available on given filters"}}}},"/assetservice/updateasset":{"put":{"tags":["Asset Services"],"summary":"Updates biopsy and fixative details for a given asset","description":"","operationId":"updateAsset","parameters":[{"name":"assetId","in":"query","description":"Asset ID","required":false,"type":"string"},{"name":"biopsy","in":"query","description":"Biopsy value","required":false,"type":"string"},{"name":"fixative","in":"query","description":"Fixative value","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/assetservice/updatespecial":{"post":{"tags":["Asset Services"],"summary":"Updates the asset to include special status tag","description":"","operationId":"updateSpecial","parameters":[{"name":"rootID","in":"query","description":"Root ID","required":false,"type":"string"},{"name":"specialType","in":"query","description":"Special Type value","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"boolean"}}}}},"/requestservice":{"post":{"tags":["Request Services"],"summary":"Saves the root","description":"","operationId":"saveRoot","parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"integer","format":"int32"},{"in":"body","name":"body","description":"Request Object","required":false,"schema":{"$ref":"#/definitions/Request"}}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Request"}},"400":{"description":"Requested NP Number cant be generated"}}}},"/requestservice/getrequeststatus":{"get":{"tags":["Request Services"],"summary":"Gets the Status of the request","description":"","operationId":"getRequestStatus","parameters":[{"name":"requestId","in":"query","description":"Request ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}},"400":{"description":"Request is not available"}}}},"/requestservice/nextroot":{"get":{"tags":["Request Services"],"summary":"Gives the next root","description":"","operationId":"getRootSuggestion","parameters":[{"name":"requestid","in":"query","description":"Request ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}},"400":{"description":"Error in request number"}}}},"/requestservice/noofrequestscompleted":{"get":{"tags":["Request Services"],"summary":"Finds the number of the requests completed","description":"","operationId":"totalRequestsCompleted","produces":["application/json"],"parameters":[{"name":"startDate","in":"query","description":"Start date","required":false,"type":"string"},{"name":"endDate","in":"query","description":"End date","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}}}}},"/requestservice/noofrequestsreceived":{"get":{"tags":["Request Services"],"summary":"Finds the number of the requests received","description":"","operationId":"totalRequestsReceived","produces":["application/json"],"parameters":[{"name":"startDate","in":"query","description":"Start date","required":false,"type":"string"},{"name":"endDate","in":"query","description":"End date","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}}}}},"/requestservice/patientdetails":{"get":{"tags":["Request Services"],"summary":"Gets the patients details for a given request ID","description":"","operationId":"getPatientDetails","parameters":[{"name":"samplerequestid","in":"query","description":"Request ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Patient"}},"400":{"description":"Patient Data is unavailable"},"404":{"description":"Request not found"}}}},"/stationservice/completed":{"get":{"tags":["Station Services"],"summary":"Returns the completed taks at a given station","description":"","operationId":"getCompletedTasks","parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"type":"object"}}}}}},"/stationservice/pendingassets":{"get":{"tags":["Station Services"],"summary":"Returns the pending assets at a given station","description":"","operationId":"getPendingTasks","parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"type":"object"}}}}}},"/stationservice/scan":{"get":{"tags":["Station Services"],"summary":"Scans the asset at a given station ID and validates","description":"","operationId":"scanAsset","produces":["application/json"],"parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"integer","format":"int32"},{"name":"assetId","in":"query","description":"Asset ID","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Asset"}},"400":{"description":"Wrong station"}}}},"/stationservice/stationstatistics":{"get":{"tags":["Station Services"],"summary":"Gets all stations statistics","description":"","operationId":"stationStatistics","parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}}}}},"/stationservice/transactions":{"get":{"tags":["Station Services"],"summary":"Returns the transactions occured at that station during the interval","description":"","operationId":"getTransactions","parameters":[{"name":"stationId","in":"query","description":"Station ID","required":false,"type":"string"},{"name":"startDate","in":"query","description":"Start date of interval","required":false,"type":"string"},{"name":"endDate","in":"query","description":"End date of interval","required":false,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"string"}},"400":{"description":"No transactions"}}}}},"definitions":{"NimhansAsset":{"type":"object","properties":{"assetId":{"$ref":"#/definitions/AssetId"},"flowId":{"$ref":"#/definitions/Flow"},"currentFlowIndex":{"type":"integer","format":"int32"},"parentId":{"$ref":"#/definitions/AssetId"},"status":{"type":"string","enum":["ENTERED","EXITED","FINISHED"]},"reqId":{"type":"string"},"remarks":{"type":"string"},"assetType":{"type":"string"},"biopsy":{"type":"string"},"fixative":{"type":"string"},"specimen":{"type":"string"},"specialType":{"type":"string","enum":["NORMAL","SPECIAL_GROSSING","SPECIAL_SECTIONING","SPECIAL_STAINING","SPECIAL"]},"childrenIds":{"type":"array","items":{"$ref":"#/definitions/AssetId"}}}},"Request":{"type":"object","properties":{"reqID":{"type":"string"},"rootID":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"requestStatus":{"type":"string"},"npBase":{"type":"string"},"patientDetails":{"$ref":"#/definitions/Patient"}}},"Patient":{"type":"object","properties":{"patientName":{"type":"string"},"age":{"type":"integer","format":"int32"},"sex":{"type":"string"},"npBase":{"type":"string"},"external":{"type":"integer","format":"int32"},"doctor":{"type":"string"},"mrd":{"type":"string"},"hospital":{"type":"string"},"departmentName":{"type":"string"},"unitName":{"type":"string"},"uniqueNpBase":{"type":"string"},"uhid":{"type":"string"}}},"Asset":{"type":"object","properties":{"assetId":{"$ref":"#/definitions/AssetId"},"flowId":{"$ref":"#/definitions/Flow"},"currentFlowIndex":{"type":"integer","format":"int32"},"parentId":{"$ref":"#/definitions/AssetId"},"status":{"type":"string","enum":["ENTERED","EXITED","FINISHED"]},"reqId":{"type":"string"},"remarks":{"type":"string"},"childrenIds":{"type":"array","items":{"$ref":"#/definitions/AssetId"}}}},"AssetId":{"type":"object","properties":{"id":{"type":"object"},"asset":{"$ref":"#/definitions/Asset"},"childrenStatus":{"type":"object","additionalProperties":{"type":"string"}}}},"Flow":{"type":"object","properties":{"flowId":{"type":"string"},"flow":{"type":"string"}}}}}